requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.2
pydantic>=2.8.2
pydantic-settings>=2.3.4
python-dotenv>=1.0.1
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
openai>=1.40.0
httpx>=0.27.0
aiofiles>=24.1.0
aiosqlite>=0.20.0
pypdf>=4.3.1
pdfplumber>=0.11.4
pdf2image>=1.17.0
pypdfium2>=4.30.0
pillow>=10.4.0
pytesseract>=0.3.10
pytest>=8.3.0
